@@ -16,47 +16,46 @@
   "main": "node_modules/expo/AppEntry.js",
   "dependencies": {
     "@expo-google-fonts/montserrat": "^0.1.0",
-    "@expo/vector-icons": "^12.0.0",
-    "@react-native-async-storage/async-storage": "~1.15.0",
-    "@react-native-community/masked-view": "0.1.10",
-    "@react-native-community/netinfo": "6.0.0",
+    "@expo/vector-icons": "^13.0.0",
+    "@react-native-async-storage/async-storage": "~1.17.3",
+    "@react-native-community/masked-view": "0.1.11",
+    "@react-native-community/netinfo": "9.3.0",
     "@react-navigation/compat": "^5.3.15",
     "@react-navigation/drawer": "^5.12.5",
     "@react-navigation/native": "^5.9.4",
     "@react-navigation/stack": "^5.14.5",
     "axios": "^0.21.2",
     "dotenv": "^10.0.0",
-    "expo": "^42.0.0",
-    "expo-app-loading": "^1.1.2",
-    "expo-application": "~3.2.0",
-    "expo-asset": "~8.3.3",
-    "expo-calendar": "~9.2.2",
-    "expo-constants": "~11.0.1",
-    "expo-device": "~3.3.0",
-    "expo-font": "~9.2.1",
-    "expo-linear-gradient": "~9.2.0",
-    "expo-updates": "~0.8.2",
+    "expo": "^46.0.0",
+    "expo-application": "~4.2.2",
+    "expo-asset": "~8.6.1",
+    "expo-calendar": "~10.3.0",
+    "expo-constants": "~13.2.4",
+    "expo-device": "~4.3.0",
+    "expo-font": "~10.2.0",
+    "expo-linear-gradient": "~11.4.0",
+    "expo-updates": "~0.14.5",
     "html-entities": "^2.3.2",
     "moment": "^2.22.2",
     "prop-types": "^15.6.2",
     "qs": "^6.10.1",
-    "react": "16.13.1",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-42.0.0.tar.gz",
-    "react-native-appearance": "~0.3.3",
+    "react": "18.0.0",
+    "react-native": "0.69.5",
     "react-native-collapsible": "^1.0.0",
-    "react-native-gesture-handler": "~1.10.2",
-    "react-native-maps": "0.28.0",
-    "react-native-reanimated": "~2.2.0",
+    "react-native-gesture-handler": "~2.5.0",
+    "react-native-maps": "0.31.1",
+    "react-native-reanimated": "~2.9.1",
     "react-native-root-siblings": "^4.1.1",
     "react-native-root-toast": "^3.0.1",
-    "react-native-safe-area-context": "3.2.0",
-    "react-native-screens": "~3.4.0",
-    "react-native-svg": "12.1.1",
+    "react-native-safe-area-context": "4.3.1",
+    "react-native-screens": "~3.15.0",
+    "react-native-svg": "12.3.0",
     "react-native-textinput-effects": "^0.6.1",
-    "react-native-webview": "11.6.2",
-    "sentry-expo": "^4.0.0",
-    "expo-background-fetch": "~9.2.2",
-    "expo-task-manager": "~9.2.2"
+    "react-native-webview": "11.23.0",
+    "sentry-expo": "~5.0.0",
+    "expo-background-fetch": "~10.3.0",
+    "expo-task-manager": "~10.3.0",
+    "expo-splash-screen": "~0.16.2"
   },
   "devDependencies": {
     "prettier": "^1.14.2"
