@@ -5,13 +5,13 @@ export default {
 	description: 'Ukit, a companion app for student of University of Bordeaux',
 	slug: 'Ukit',
 	privacy: 'public',
-	githubUrl: 'https://github.com/Jack3113/UKit',
+	githubUrl: 'https://github.com/kb-dev/UKit',
 	platforms: ['ios', 'android'],
-	version: '4.0.1',
+	version: '4.0.4',
 	orientation: 'portrait',
 	primaryColor: '#006F9F',
 	icon: './assets/icons/ios.png',
-	owner: 'hackjack',
+	owner: 'kbdev',
 	splash: {
 		image: './assets/icons/splash.png',
 		backgroundColor: '#009de0',
